add confirm delete window add update user function

+++ app/userPage/usersTableCtrl.js
@@ -25,8 +25,10 @@ app.controller("usersTableCtrl", function ($scope, $location, $uibModal, userSrv
                     return {
                         fname: user.fname,
                         lname: user.lname,
-                        email: user.email,
-                        phone: user.phone
+                        email: user.copyOfEmail,
+                        phone: user.phone,
+                        isOtherUser: true,
+                        userId: user.userId
                     };
                 }
             }
@@ -42,13 +44,32 @@ app.controller("usersTableCtrl", function ($scope, $location, $uibModal, userSrv
     }
 
     $scope.goToDeleteUser = function (user) {
-        userSrv.deleteUser(user).then(function (response) {
-           console.log(response);
-        });
-        /**update table after deleting a user */
-        userSrv.getAllUsers().then(function (users) {
-            $scope.users = users;
-        });
+        /** open the delete slide modal */
+        var modalInstance = $uibModal.open({
+            templateUrl: "app/signUpModal/userDeleteModal.html",
+            controller: "userDeleteModalCtrl",
+            resolve: {
+                userToDel: function () {
+                    return user;
+                }
+            }
+        })
+        modalInstance.result.then(function (deleteWasConfirmed) {
+            // this will wake in case the user deleted the current slide
+            if (deleteWasConfirmed) {
+                userSrv.deleteUser(user).then(function (response) {
+                    console.log(response);
+                    /**update table after deleting a user */
+                    userSrv.getAllUsers().then(function (users) {
+                        $scope.users = users;
+                    });
+                });
+
+            }
+        }, function () {
+            // this will wake up in case the user canceled the update
+            console.log("user canceled delete");
+        })
     }
 
 });
